AURORA_ENGINE/requirements-aurora.txt
notebooklm-py[browser]
playwright
httpx
rich
mempalace
chromadb
pysqlite3-binary
